+++ mazikeen/MakedirsBlock.py
@@ -8,6 +8,7 @@ class MakedirsBlock:
         self.dir = dir
 
     def run(self, workingDir = "", variables = {}, printer = Printer()):
+        printer.verbose("Makedirs:", self.dir)
         _dir = replaceVariables(self.dir, variables, printer)
         _dir = str(pathlib.PurePath(workingDir).joinpath(_dir)) + "/" 
         ensure_dir(_dir)