+++ changelog.rst
@@ -3,16 +3,21 @@ For a complete changelog, see:
 
 * https://github.com/hanniballar/mazikeen/commits
 
+1.1.0
+-----
+* Diff block replaced attributes `leftpath` and `rightpath` with paths. Script version 1.1.0 created
+* Run block without defined, or None exitcode will ignore exit code.
+* Add the name of every block as verbose output.
+
 1.0.2
 -----
-* Run block replaces variables for attributes 'outputfile' and 'inputfile'
-* Run block without defined, or None exitcode will ignore exit code.
+* Run block replaces variables for attributes 'outputfile' and 'inputfile'.
 
 1.0.1
 -----
 * Run block didn't work properly on linux.
 * Test case wasn't marked as failed if it failed because of diff block.
-* added '--version' command line argument
+* added '--version' command line argument.
 
 1.0.0
 -----